src/main/java/com/example/demo/service/StudentServiceImpl.java
package com.example.demo.service;

import com.example.demo.dao.StudentDao;
import com.example.demo.model.Student;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class StudentServiceImpl implements StudentService {

    private final StudentDao studentDao;

    public StudentServiceImpl(StudentDao studentDao) {
        this.studentDao = studentDao;
    }

    @Override
    public Student save(Student student) {
        if (student == null) {
            throw new IllegalArgumentException("Student cannot be null!");
        }
        try {
            return studentDao.save(student);
        } catch (Exception e) {
            e.printStackTrace();
            return null;
        }
    }

    @Override
    public List<Student> findAll() {
        try {
            return studentDao.findAll();
        } catch (Exception e) {
            e.printStackTrace();
            return null;
        }
    }

    @Override
    public List<Student> findStudentsByCourseName(String courseName) {
        try {
            return studentDao.findStudentsByCourseName(courseName);
        } catch (Exception e) {
            e.printStackTrace();
            return null;
        }
    }

    @Override
    public Student findById(Long student_id) {
        try {
            return studentDao.findById(student_id);
        } catch (Exception e) {
            e.printStackTrace();
            return null;
        }
    }

    @Override
    public void deleteById(Long student_id) {
        try{
            studentDao.deleteById(student_id);
        } catch (Exception e){
            e.printStackTrace();
        }
    }
}
